Move RazorIndentationFacts to LegacyEditor project and reorganize tests
