+++ src/Razor/src/Microsoft.VisualStudio.LegacyEditor.Razor/Indentation/RazorIndentationFacts.cs
@@ -8,7 +8,7 @@ using Microsoft.AspNetCore.Razor.Language.Legacy;
 using Microsoft.AspNetCore.Razor.Language.Syntax;
 using Microsoft.VisualStudio.Text;
 
-namespace Microsoft.VisualStudio.Editor.Razor;
+namespace Microsoft.VisualStudio.LegacyEditor.Razor.Indentation;
 
 internal static class RazorIndentationFacts
 {
