+++ src/Razor/test/Microsoft.VisualStudio.LegacyEditor.Razor.Test/Completion/VisualStudioDescriptionFactoryTest.cs
@@ -9,7 +9,7 @@ using Microsoft.VisualStudio.Text.Adornments;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Microsoft.VisualStudio.LegacyEditor.Razor.Completion.Test;
+namespace Microsoft.VisualStudio.LegacyEditor.Razor.Completion;
 
 public class VisualStudioDescriptionFactoryTest(ITestOutputHelper testOutput) : ToolingTestBase(testOutput)
 {
