+++ src/Razor/test/Microsoft.VisualStudio.LegacyEditor.Razor.Test/Indentation/BraceSmartIndenterTestBase.cs
@@ -13,7 +13,7 @@ using Microsoft.VisualStudio.Text.Operations;
 using Moq;
 using Xunit.Abstractions;
 
-namespace Microsoft.VisualStudio.LegacyEditor.Razor.Test;
+namespace Microsoft.VisualStudio.LegacyEditor.Razor.Indentation;
 
 public class BraceSmartIndenterTestBase(ITestOutputHelper testOutput) : ProjectSnapshotManagerDispatcherTestBase(testOutput)
 {
