+++ src/Razor/test/Microsoft.VisualStudio.LegacyEditor.Razor.Test/Microsoft.VisualStudio.LegacyEditor.Razor.Test.csproj
@@ -3,10 +3,18 @@
   <PropertyGroup>
     <TargetFramework>$(DefaultNetFxTargetFramework)</TargetFramework>
     <PreserveCompilationContext>true</PreserveCompilationContext>
+    <RootNamespace>Microsoft.VisualStudio.LegacyEditor.Razor</RootNamespace>
   </PropertyGroup>
 
   <ItemGroup>
-    <EmbeddedResource Include="TestFiles\**\*" />
+    <EmbeddedResource Include="TestFiles\**\*">
+      <!-- This ensures that the expected resource name is generated for test files.
+           Normally, the resource name would be based on the root namespace, but this
+           project changes $(RootNamespace) above. Because %(RecursiveDir) has slashes
+           in it, we need to turn those into dots to get the final resource name to be
+           what the test infrastructure expects.-->
+      <LogicalName>$(ProjectName).TestFiles.$([MSBuild]::ValueOrDefault('%(RecursiveDir)', '').Replace('\', '.').Replace('/', '.'))%(Filename)%(Extension)</LogicalName>
+    </EmbeddedResource>
     <None Include="xunit.runner.json" CopyToOutputDirectory="PreserveNewest" />
   </ItemGroup>
