+++ src/Razor/test/Microsoft.VisualStudio.LegacyEditor.Razor.Test/Parsing/VisualStudioRazorParserTest.cs
@@ -11,12 +11,11 @@ using Microsoft.AspNetCore.Razor.Test.Common.Workspaces;
 using Microsoft.CodeAnalysis.Razor.ProjectSystem;
 using Microsoft.VisualStudio.Editor.Razor;
 using Microsoft.VisualStudio.Language.Intellisense;
-using Microsoft.VisualStudio.LegacyEditor.Razor.Parsing;
 using Moq;
 using Xunit;
 using Xunit.Abstractions;
 
-namespace Microsoft.VisualStudio.LegacyEditor.Razor.Parsing.Test;
+namespace Microsoft.VisualStudio.LegacyEditor.Razor.Parsing;
 
 public class VisualStudioRazorParserTest : ProjectSnapshotManagerDispatcherTestBase
 {
